Remove backend/utils/helpers.js
